move core/types/accounts to erigon-lib/types/accounts
