+++ cmd/verkle/verkletrie/workers.go
@@ -22,7 +22,7 @@ import (
 	libcommon "github.com/erigontech/erigon-lib/common"
 	"github.com/holiman/uint256"
 
-	"github.com/erigontech/erigon/core/types/accounts"
+	"github.com/erigontech/erigon-lib/types/accounts"
 	"github.com/erigontech/erigon/turbo/trie/vtree"
 )
