+++ turbo/shards/state_cache.go
@@ -29,7 +29,7 @@ import (
 	libcommon "github.com/erigontech/erigon-lib/common"
 	"github.com/erigontech/erigon-lib/metrics"
 
-	"github.com/erigontech/erigon/core/types/accounts"
+	"github.com/erigontech/erigon-lib/types/accounts"
 )
 
 // LRU state cache consists of two structures - B-Tree and binary heap
