+++ turbo/trie/delete_subrtee_test.go
@@ -28,7 +28,7 @@ import (
 	"github.com/erigontech/erigon-lib/common/length"
 	"github.com/erigontech/erigon-lib/crypto"
 
-	"github.com/erigontech/erigon/core/types/accounts"
+	"github.com/erigontech/erigon-lib/types/accounts"
 )
 
 func TestTrieDeleteSubtree_ShortNode(t *testing.T) {
