+++ turbo/trie/retain_list_test.go
@@ -24,7 +24,7 @@ import (
 
 	libcommon "github.com/erigontech/erigon-lib/common"
 	"github.com/erigontech/erigon-lib/common/hexutil"
-	"github.com/erigontech/erigon/core/types/accounts"
+	"github.com/erigontech/erigon-lib/types/accounts"
 )
 
 func FakePreimage(hash libcommon.Hash) libcommon.Hash {
