+++ turbo/trie/stream_test.go
@@ -28,7 +28,7 @@ import (
 	"github.com/erigontech/erigon-lib/crypto"
 
 	"github.com/erigontech/erigon-lib/common"
-	"github.com/erigontech/erigon/core/types/accounts"
+	"github.com/erigontech/erigon-lib/types/accounts"
 )
 
 func TestHashWithModificationsEmpty(t *testing.T) {
