+++ turbo/trie/trie_transform.go
@@ -18,7 +18,7 @@ package trie
 
 import (
 	"github.com/erigontech/erigon-lib/common"
-	"github.com/erigontech/erigon/core/types/accounts"
+	"github.com/erigontech/erigon-lib/types/accounts"
 )
 
 type keyTransformFunc func([]byte) []byte
